Peripheral API v1.4.0 - Offset discrete D-pads
Discrete D-pads are D-pads that use analog axes with discrete values
instead of digital buttons.

Normally, discrete D-pads are centered at 0.0, travel to -1.0 when left is
pressed, and 1.0 when right is pressed.

On Windows, discrete D-pads have been observed that start centered at 0.498,
travel to 0.0 when left is pressed, and 1.0 when right is pressed.

+++ xbmc/addons/kodi-addon-dev-kit/include/kodi/addon-instance/Peripheral.h
@@ -181,9 +181,9 @@ extern "C"
   typedef struct JOYSTICK_DRIVER_SEMIAXIS
   {
     int                                index;
-    int                                center;
+    float                              center;
     JOYSTICK_DRIVER_SEMIAXIS_DIRECTION direction;
-    unsigned int                       range;
+    float                              range;
   } ATTRIBUTE_PACKED JOYSTICK_DRIVER_SEMIAXIS;
 
   typedef struct JOYSTICK_DRIVER_MOTOR
