+++ xbmc/input/joysticks/generic/InputHandling.h
@@ -57,7 +57,7 @@ namespace JOYSTICK
     // implementation of IDriverHandler
     virtual bool OnButtonMotion(unsigned int buttonIndex, bool bPressed) override;
     virtual bool OnHatMotion(unsigned int hatIndex, HAT_STATE state) override;
-    virtual bool OnAxisMotion(unsigned int axisIndex, float position, int center, unsigned int range) override;
+    virtual bool OnAxisMotion(unsigned int axisIndex, float position, float center, float range) override;
     virtual void ProcessAxisMotions(void) override;
 
   private:
