+++ xbmc/peripherals/addons/AddonButtonMapping.cpp
@@ -76,7 +76,7 @@ bool CAddonButtonMapping::OnHatMotion(unsigned int hatIndex, HAT_STATE state)
   return false;
 }
 
-bool CAddonButtonMapping::OnAxisMotion(unsigned int axisIndex, float position, int center, unsigned int range)
+bool CAddonButtonMapping::OnAxisMotion(unsigned int axisIndex, float position, float center, float range)
 {
   if (m_buttonMapping)
     return m_buttonMapping->OnAxisMotion(axisIndex, position, center, range);
